exclude development commands from deliverables
The following warnings occur for charwidth v0.2.0 and rubocop or rspec users. This is because the charwidth artifacts include executable files for development, such as rubocop and rspec.

```
The `rubocop` executable in the `charwidth` gem is being loaded, but it's also present in other gems (rubocop).
If you meant to run the executable for another gem, make sure you use a project specific binstub (`bundle binstub <gem_name>`).
If you plan to use multiple conflicting executables, generate binstubs for them and disambiguate their names.
The `rubocop` executable in the `rubocop` gem is being loaded, but it's also present in other gems (charwidth).
If you meant to run the executable for another gem, make sure you use a project specific binstub (`bundle binstub <gem_name>`).
If you plan to use multiple conflicting executables, generate binstubs for them and disambiguate their names.
```